Little refactoring.
